Solves problem with second-level dependencies (#22)
Related to https://github.com/insightsengineering/verdepcheck/issues/38

In scenarios where the `DESCRIPTION` file has top-level dependencies
(`Depends`, `Imports`, `Suggests`, `Extends`) that depend on packages
not yet released on CRAN, this new option allows to define repositories
that may fill the gap.

In particular, this is important on `max` strategy, as they might depend
on secondary packages that have changes that only exist on the
repository.

### Changes description

- Adds new option to add extra repositories
- If user wants to fully override default repository, they can replace
environmental variables
- Adds styling to first (manual) header
- Better organizes readme

### TODO:

- [ ] Understand how to use this new option in github action `matrix`

### Use case examples

This will be especially helpful with `max` strategy to allow more
packages to be compiled with latest development versions

- `insightsengineering` packages can use `r-universe.dev` to pull
secondary dependencies such as `formatters`
- Pull all latest development from tidyverse packages by using
https://tidyverse.r-universe.dev/builds

`release` strategy can also use this if downstream packages are not
being release on CRAN

`min_***` strategy shouldn't use this option